BitmapImage::setFullyOpaque() and small fixes

+++ src/common_defines.hpp
@@ -6,13 +6,13 @@
 typedef char int8;
 typedef short int16;
 typedef int int32;
-typedef long int64;
+typedef long long int64;
 typedef bool bool32;
 
 typedef unsigned char uint8;
 typedef unsigned short uint16;
 typedef unsigned int uint32;
-typedef unsigned long uint64;
+typedef unsigned long long uint64;
 
 // TEMPORARY: debug only, to fix vscode highlighting
 #ifndef DEBUG
